Debugging

+++ Gbox420_Nano_Hempy/src/Components/Switch.cpp
@@ -15,7 +15,7 @@ void Switch::report(bool JSONReport)
   if (JSONReport) //Caller requested a JSON formatted report: Append it to the LogMessage buffer. Caller is responsible of clearing the LongMessage buffer
   {
     strcat_P(LongMessage, (PGM_P)F("\"S\":\""));
-    strcat(LongMessage, toText((int)getState()));
+    strcat(LongMessage, toText(getState()));
     strcat_P(LongMessage, (PGM_P)F("\"}")); ///< closing the curly bracket at the end of the JSON
   }
   else //Print a report to the Serial console
